moving config related methods in llmisvc controller to separate file
