+++ pkg/controller/v1alpha1/llmisvc/controller.go
@@ -18,16 +18,10 @@ package llmisvc
 
 import (
 	"context"
-	"fmt"
-	"strings"
 
 	"k8s.io/client-go/kubernetes"
 
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
-
-	"github.com/kserve/kserve/pkg/apis/serving/v1beta1"
-	"github.com/kserve/kserve/pkg/constants"
-	kserveTypes "github.com/kserve/kserve/pkg/types"
 
 	"k8s.io/client-go/tools/record"
 
@@ -36,15 +30,6 @@ import (
 
 	"github.com/kserve/kserve/pkg/apis/serving/v1alpha1"
 )
-
-type Config struct {
-	SystemNamespace             string   `json:"systemNamespace,omitempty"`
-	IngressGatewayName          string   `json:"ingressGatewayName,omitempty"`
-	IngressGatewayNamespace     string   `json:"ingressGatewayNamespace,omitempty"`
-	IstioGatewayControllerNames []string `json:"istioGatewayControllerNames,omitempty"`
-
-	StorageConfig *kserveTypes.StorageInitializerConfig `json:"-"`
-}
 
 // +kubebuilder:rbac:groups=serving.kserve.io,resources=llminferenceservices;llminferenceservices/finalizers,verbs=get;list;watch;create;update;patch;delete
 // +kubebuilder:rbac:groups=serving.kserve.io,resources=llminferenceservices/status,verbs=get;update;patch
@@ -55,50 +40,6 @@ type LLMISVCReconciler struct {
 	client.Client
 	record.EventRecorder
 	Clientset kubernetes.Interface
-}
-
-// NewConfig creates an instance of llm-specific config based on predefined values
-// in IngressConfig struct
-func NewConfig(ingressConfig *v1beta1.IngressConfig, storageConfig *kserveTypes.StorageInitializerConfig) *Config {
-	igwNs := constants.KServeNamespace
-	igwName := ingressConfig.KserveIngressGateway
-	igw := strings.Split(igwName, "/")
-	if len(igw) == 2 {
-		igwNs = igw[0]
-		igwName = igw[1]
-	}
-
-	return &Config{
-		SystemNamespace:         constants.KServeNamespace,
-		IngressGatewayNamespace: igwNs,
-		IngressGatewayName:      igwName,
-		// TODO make it configurable
-		IstioGatewayControllerNames: []string{
-			"istio.io/gateway-controller",
-			"istio.io/unmanaged-gateway",
-			"openshift.io/gateway-controller",
-		},
-		StorageConfig: storageConfig,
-	}
-}
-
-func LoadConfig(ctx context.Context, clientset kubernetes.Interface) (*Config, error) {
-	isvcConfigMap, errCfgMap := v1beta1.GetInferenceServiceConfigMap(ctx, clientset) // Fetch directly from API Server
-	if errCfgMap != nil {
-		return nil, fmt.Errorf("failed to load InferenceServiceConfigMap: %w", errCfgMap)
-	}
-
-	ingressConfig, errConvert := v1beta1.NewIngressConfig(isvcConfigMap)
-	if errConvert != nil {
-		return nil, fmt.Errorf("failed to convert InferenceServiceConfigMap to IngressConfig: %w", errConvert)
-	}
-
-	storageInitializerConfig, errConvert := v1beta1.GetStorageInitializerConfigs(isvcConfigMap)
-	if errConvert != nil {
-		return nil, fmt.Errorf("failed to convert InferenceServiceConfigMap to StorageInitializerConfig: %w", errConvert)
-	}
-
-	return NewConfig(ingressConfig, storageInitializerConfig), nil
 }
 
 func (r *LLMISVCReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
